Utilities/Drive/requirements.txt
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
google-api-core==2.1.1
google-api-python-client==2.26.1
google-auth==2.3.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
httplib2==0.20.1
idna==3.3
oauthlib==3.1.1
protobuf==3.18.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.7
